Create book repository

+++ internal/catalog/model/paginated_books.go
@@ -4,5 +4,5 @@ type PaginatedBooks struct {
 	Books      []Book
 	Limit      int
 	Offset     int
-	TotalBooks int
+	TotalBooks int64
 }
